fix: 表示される緯度経度を整数にする

+++ app/src/main/java/com/example/snap_develop/view/adapter/TimelineAdapter.java
@@ -71,7 +71,7 @@ public class TimelineAdapter extends RecyclerView.Adapter<TimelineViewHolder> {
         if (TextUtils.equals(postBean.getType(), "post")) {
             holder.mGoodCountTextView.setText(postBean.getGoodCount());
             holder.mLatLngTextView.setText(
-                    String.format("%s, %s", postBean.getLatLng().latitude, postBean.getLatLng().longitude));
+                    String.format("%s, %s", (int) postBean.getLatLng().latitude, (int) postBean.getLatLng().longitude));
         }
 
         holder.mUserInfoConstraintLayout.setOnClickListener(new View.OnClickListener() {
